requirements.txt
# Core dependencies
linked-claims-extractor>=0.1.5
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
gunicorn>=20.0.0

# Database
psycopg2-binary>=2.9.9

# Background processing
celery>=5.3.0
redis>=5.0.0

# PDF processing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
pillow>=11.0.0

# API dependencies
requests>=2.32.0
python-dotenv>=1.0.0

# AI/ML dependencies - REQUIRED for claim extraction
anthropic>=0.39.0
langchain>=0.3.0
langchain-anthropic>=0.3.0

# Optional - only if using OpenAI instead
# openai>=1.55.0
